requirements.txt
Flask==1.0.2
Flask-Cors==3.0.8
PyJWT==1.7.1
SQLAlchemy==1.2.7